♻️ send json to route api instead of file
