+++ apps/web/package.json
@@ -59,7 +59,6 @@
     "dayjs": "^1.11.7",
     "decimal.js": "^10.4.3",
     "drizzle-orm": "^0.30.7",
-    "file-type": "^19.0.0",
     "lucide-react": "^0.259.0",
     "next": "^14.2.0",
     "react": "^18.2.0",
